+++ core/index.ts
@@ -80,6 +80,7 @@ export type {
   LytxAiModel,
 } from "./src/config/createLytxAppConfig";
 export type {
+  LytxAdditionalRoute,
   LytxDashboardReportData,
   LytxDashboardRouteUiOverrideArgs,
   LytxEventsRouteUiOverrideArgs,